clean up based on design feedback
