+++ src/compare/Compare.css
@@ -27,7 +27,3 @@ body {
 .right-panel {
   margin-top: 40px;
 }
-
-.zoom-overlay .OverlayContainer {
-  display: block;
-};
